# annotrieve-api-specs.yaml
openapi: 3.1.0
info:
  version: "1.0.0"
  title: "Annotrieve API"
  description: "API for retrieving and managing eukaryotic genome annotations"
  contact:
    email: "emilio.righi@crg.eu"
servers:
  - url: https://genome.crg.es/annotrieve/api/v0
    description: Production server
...
paths:
  /annotations:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotations"
      summary: "Get annotations"
      description: "Returns annotation metadata filtered, sorted, and paginated via query parameters."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
        - $ref: "#/components/parameters/fields"
      responses:
        "200":
          description: "Annotations page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotations"
      summary: "Get annotations via POST"
      description: "Same as GET /annotations, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Annotations page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/report:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationsReport"
      summary: "Download annotation metadata as TSV"
      description: "Streams a TSV file with annotation metadata columns (annotation_id, assembly_accession, assembly_name, organism_name, taxid, database, provider, source_url, bgzip_path, csi_path)."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "TSV payload streamed line by line"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotationsReport"
      summary: "Download annotation metadata via POST"
      description: "Same as GET /annotations/report, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "TSV payload streamed line by line"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/frequencies:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationFrequencyFields"
      summary: "List allowed annotation frequency fields"
      responses:
        "200":
          description: "Frequency-capable fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyFieldList"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/frequencies/{field}:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationFrequencies"
      summary: "Get annotation field frequencies"
      description: "Returns frequency counts for the selected field across filtered annotations."
      parameters:
        - name: field
          in: path
          required: true
          description: "Field to aggregate. Allowed values: taxid, organism_name, database, release_date, assembly_accession, assembly_name, feature_type, feature_source, biotype, pipeline, provider."
          schema:
            type: string
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postAnnotationFrequencies"
      summary: "Get annotation field frequencies via POST"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/errors:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationErrors"
      summary: "List annotation import errors"
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: "Annotation errors"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAnnotationErrorsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/gene-stats:
    get:
      tags:
        - "annotations"
      operationId: "getGeneStats"
      summary: "Get gene stats summary"
      description: "Returns gene stats summary with aggregated statistics across all categories (coding, non_coding, pseudogene)."
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
      responses:
        "200":
          description: "Gene stats summary"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneStatsSummaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postGeneStats"
      summary: "Get gene stats summary via POST"
      description: "Same as GET /annotations/gene-stats, but accepts filters in the request body."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Gene stats summary"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneStatsSummaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/gene-stats/{category}:
    get:
      tags:
        - "annotations"
      operationId: "getGeneCategoryDetails"
      summary: "Get detailed statistics for a specific gene category"
      description: "Returns detailed statistics for a specific gene category (coding, non_coding, or pseudogene)."
      parameters:
        - name: category
          in: path
          required: true
          description: "Gene category name (coding, non_coding, or pseudogene)"
          schema:
            type: string
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
      responses:
        "200":
          description: "Gene category details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneCategoryDetailsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postGeneCategoryDetails"
      summary: "Get gene category details via POST"
      description: "Same as GET /annotations/gene-stats/{category}, but accepts filters in the request body."
      parameters:
        - name: category
          in: path
          required: true
          description: "Gene category name (coding, non_coding, or pseudogene)"
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationQueryParams"
      responses:
        "200":
          description: "Gene category details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneCategoryDetailsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /annotations/gene-stats/{category}/{metric}:
    get:
      tags:
        - "annotations"
      operationId: "getGeneCategoryMetricValues"
      summary: "Get raw values for a specific metric in a gene category"
      description: "Returns raw values for a specific metric in a specific gene category (for plotting histograms)."
      parameters:
        - name: category
          in: path
          required: true
          description: "Gene category name (coding, non_coding, or pseudogene)"
          schema:
            type: string
        - name: metric
          in: path
          required: true
          description: "Metric name (total_count or average_mean_length)"
          schema:
            type: string
        - name: include_annotations
          in: query
          description: "If true, include annotation_ids list in response"
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/bioproject_accessions"
        - $ref: "#/components/parameters/db_sources"
        - $ref: "#/components/parameters/feature_sources"
        - $ref: "#/components/parameters/biotypes"
        - $ref: "#/components/parameters/feature_types"
        - $ref: "#/components/parameters/pipelines"
        - $ref: "#/components/parameters/providers"
        - $ref: "#/components/parameters/md5_checksums"
        - $ref: "#/components/parameters/has_stats"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/release_date_from"
        - $ref: "#/components/parameters/release_date_to"
      responses:
        "200":
          description: "Metric values"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneCategoryMetricValuesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "annotations"
      operationId: "postGeneCategoryMetricValues"
      summary: "Get gene category metric values via POST"
      description: "Same as GET /annotations/gene-stats/{category}/{metric}, but accepts filters in the request body."
      parameters:
        - name: category
          in: path
          required: true
          description: "Gene category name (coding, non_coding, or pseudogene)"
          schema:
            type: string
        - name: metric
          in: path
          required: true
          description: "Metric name (total_count or average_mean_length)"
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/AnnotationQueryParams"
                - type: object
                  properties:
                    include_annotations:
                      type: boolean
                      description: "If true, include annotation_ids list in response"
                      default: false
      responses:
        "200":
          description: "Metric values"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneCategoryMetricValuesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /annotations/{md5_checksum}:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationById"
      summary: "Get annotation by checksum"
      parameters:
        - name: md5_checksum
          in: path
          required: true
          description: "MD5 checksum of the annotation (32 lowercase hex characters)."
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
      responses:
        "200":
          description: "Annotation details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Annotation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /annotations/{md5_checksum}/contigs:
    get:
      tags:
        - "annotations"
      operationId: "getAnnotationContigs"
      summary: "List contigs available for an annotation"
      description: "Streams newline-separated contig identifiers from the annotation's Tabix index."
      parameters:
        - name: md5_checksum
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
      responses:
        "200":
          description: "Contig IDs"
          content:
            text/plain:
              schema:
                type: string
                description: "Newline-separated contig identifiers"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    filter:
      name: "filter"
      in: "query"
      description: "Text search filter across multiple fields"
      schema:
        type: "string"
    limit:
      name: "limit"
      in: "query"
      description: "Maximum number of results to return"
      schema:
        type: "integer"
        minimum: 1
        maximum: 1000
        default: 20
    offset:
      name: "offset"
      in: "query"
      description: "Number of results to skip"
      schema:
        type: "integer"
        minimum: 0
        default: 0
    taxids:
      name: "taxids"
      in: "query"
      description: "Comma-separated list of taxonomy IDs"
      schema:
        type: "string"
      example: "9606,10090,7227"
    bioproject_accessions:
      name: "bioproject_accessions"
      in: "query"
      description: "Comma-separated list of BioProject accessions"
      schema:
        type: "string"
      example: "PRJNA123456,PRJNA123457"
    assembly_accessions:
      name: "assembly_accessions"
      in: "query"
      description: "Comma-separated list of assembly accessions"
      schema:
        type: "string"
      example: "GCA_000001405.29,GCF_000001405.40"
    db_sources:
      name: "db_sources"
      in: "query"
      description: "Comma-separated list of source databases (RefSeq, Ensembl, GenBank)"
      schema:
        type: "string"
    feature_sources:
      name: "feature_sources"
      in: "query"
      description: "Comma-separated list of GFF feature sources (column 2)"
      schema:
        type: "string"
    biotypes:
      name: "biotypes"
      in: "query"
      description: "Comma-separated list of feature biotypes"
      schema:
        type: "string"
    feature_types:
      name: "feature_types"
      in: "query"
      description: "Comma-separated list of GFF feature types (column 3)"
      schema:
        type: "string"
...
  schemas:
...
    FrequencyCounts:
      type: object
      additionalProperties:
        type: integer
...